+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 ## 4.13.0 [unreleased]
 
 ### Features
-1. [#528](https://github.com/influxdata/influxdb-client-csharp/pull/528): Add HttpClient as a part of InfluxDBClientOptions
+  1. [#528](https://github.com/influxdata/influxdb-client-csharp/pull/528): Add HttpClient as a part of InfluxDBClientOptions
+  1. [#500](https://github.com/influxdata/influxdb-client-csharp/pull/500): Support String functions in Linq
 
 ### Dependencies
 Update dependencies: